// Game-Hub/game-Hub-fe/src/app/services/models/page-response-user-library-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { UserLibraryResponse } from '../models/user-library-response';
export interface PageResponseUserLibraryResponse {
  content?: Array<UserLibraryResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}